OGM-359 separate test IndexedStoredProcedureCallTest
The PR contans versions of the test for 'core' module and other
dialects.
